+++ src/style.css
@@ -1,182 +1,176 @@
-@font-face {
-  font-family: "StratosSkyeng";
-  src: url("../fonts/stratosskyengweb-regular.woff") format("woff"), url("../fonts/stratosskyengweb-regular.woff2") format("woff2");
-  font-weight: normal;
-  font-style: normal;
-}
 * {
-  padding: 0px;
-  margin: 0px;
+    padding: 0px;
+    margin: 0px;
 }
 
 .conteiner {
-  height: 100vh;
-  background-color: #004980;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+    height: 100vh;
+    background-color: #004980;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
 }
 
 .content__screen1 {
-  width: 480px;
-  height: 459px;
-  background-color: #c2f5ff;
-  border-radius: 12px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  align-items: center;
+    width: 480px;
+    height: 459px;
+    background-color: #c2f5ff;
+    border-radius: 12px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
 }
 .content__screen2 {
-  padding: 25px;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  width: 950px;
+    padding: 25px;
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-end;
+    width: 950px;
 }
 .content__loss-win {
-  width: 480px;
-  height: 459px;
-  background: #c2f5ff;
-  border-radius: 12px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
+    width: 480px;
+    height: 459px;
+    background: #c2f5ff;
+    border-radius: 12px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 10px;
 }
 
 .title {
-  width: 208px;
-  height: 96px;
-  font-size: 40px;
-  line-height: 48px;
-  font-family: "StratosSkyeng";
-  font-style: normal;
-  text-align: center;
-  color: #004980;
+    width: 208px;
+    height: 96px;
+    font-size: 40px;
+    line-height: 48px;
+    font-family: 'StratosSkyeng';
+    font-style: normal;
+    text-align: center;
+    color: #004980;
 }
 .title__loss {
-  width: 276px;
+    width: 276px;
 }
 
 .button__lvl {
-  display: flex;
-  gap: 24px;
+    display: flex;
+    gap: 24px;
 }
 
 .btn {
-  background: #ffffff;
-  border-radius: 12px;
-  width: 98px;
-  height: 98px;
-  font-family: "StratosSkyeng";
-  font-size: 64px;
-  line-height: 72px;
-  color: #0080c1;
-  border-color: transparent;
+    background: #ffffff;
+    border-radius: 12px;
+    width: 98px;
+    height: 98px;
+    font-family: 'StratosSkyeng';
+    font-size: 64px;
+    line-height: 72px;
+    color: #0080c1;
+    border-color: transparent;
 }
 
 .button {
-  background: #7ac100;
-  border-radius: 12px;
-  width: 246px;
-  height: 50px;
-  border-color: transparent;
-  font-family: "StratosSkyeng";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 24px;
-  line-height: 32px;
-  color: #ffffff;
+    background: #7ac100;
+    border-radius: 12px;
+    width: 246px;
+    height: 50px;
+    border-color: transparent;
+    font-family: 'StratosSkyeng';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 24px;
+    line-height: 32px;
+    color: #ffffff;
 }
 .button__hiding {
-  background-color: gray;
+    background-color: gray;
 }
 
 .add__color {
-  background-color: #7ac100;
+    background-color: #7ac100;
 }
 
 .min-sec {
-  position: relative;
-  display: flex;
-  justify-content: flex-end;
-  top: 10px;
-  padding-right: 7px;
-  font-family: "StratosSkyeng";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 32px;
-  color: #ffffff;
+    position: relative;
+    display: flex;
+    justify-content: flex-end;
+    top: 10px;
+    padding-right: 7px;
+    font-family: 'StratosSkyeng';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 32px;
+    color: #ffffff;
 }
 
 .min {
-  padding-right: 55px;
+    padding-right: 55px;
 }
 
 .watch {
-  display: flex;
-  font-family: "StratosSkyeng";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 64px;
-  line-height: 72px;
-  color: #ffffff;
+    display: flex;
+    font-family: 'StratosSkyeng';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 64px;
+    line-height: 72px;
+    color: #ffffff;
 }
 .watch__loss {
-  color: #000000;
-  margin-bottom: 30px;
+    color: #000000;
+    margin-bottom: 30px;
 }
 
 .card-field {
-  display: flex;
-  width: 842px;
-  height: 543px;
-  flex-wrap: wrap;
-  gap: 30px;
-  justify-content: center;
-  align-items: center;
-  align-content: center;
+    display: flex;
+    width: 842px;
+    height: 543px;
+    flex-wrap: wrap;
+    gap: 30px;
+    justify-content: center;
+    align-items: center;
+    align-content: center;
 }
 
 .card {
-  position: relative;
-  z-index: 1;
+    position: relative;
+    z-index: 1;
 }
 
 .box {
-  width: 95px;
-  height: 133px;
+    width: 95px;
+    height: 133px;
 }
 
 .shirt {
-  width: 83px;
-  height: 121px;
-  background: #e4fbff;
-  border-radius: 4px;
-  padding: 6px;
-  position: relative;
-  bottom: 133px;
+    width: 83px;
+    height: 121px;
+    background: #e4fbff;
+    border-radius: 4px;
+    padding: 6px;
+    position: relative;
+    bottom: 133px;
 }
 
 .shirt_pick {
-  z-index: 2;
+    z-index: 2;
 }
 
 .emoji {
-  width: 90px;
-  height: 90px;
-  margin-top: 22px;
+    width: 90px;
+    height: 90px;
+    margin-top: 22px;
 }
 
 .paragraph {
-  font-family: "StratosSkyeng";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 24px;
-  line-height: 32px;
-  color: #000000;
-}/*# sourceMappingURL=style.css.map */
+    font-family: 'StratosSkyeng';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 24px;
+    line-height: 32px;
+    color: #000000;
+} /*# sourceMappingURL=style.css.map */
